The Swift version is now working perfectly.  Tidy up time.
