+++ ParticleEmitterDemo-SK-Swift/GameViewController.swift
@@ -16,19 +16,18 @@ class GameViewController: UIViewController {
         super.viewDidLoad()
         
         if let view = self.view as! SKView? {
-            // Load the SKScene from 'GameScene.sks'
-            if let scene = SKScene(fileNamed: "GameScene") {
-                // Set the scale mode to scale to fit the window
-                scene.scaleMode = .aspectFill
-                
-                // Present the scene
-                view.presentScene(scene)
-            }
-            
-            view.ignoresSiblingOrder = true
-            
             view.showsFPS = true
             view.showsNodeCount = true
+
+            /* Sprite Kit applies additional optimizations to improve rendering performance */
+            view.ignoresSiblingOrder = true
+
+            // Load the SKScene from 'GameScene.sks'
+            let scene = GameScene(size: CGSize(width: 320, height: 568))
+                //SKScene(fileNamed: "GameScene")
+                
+            // Present the scene
+            view.presentScene(scene)
         }
     }
 
@@ -47,4 +46,5 @@ class GameViewController: UIViewController {
     override var prefersStatusBarHidden: Bool {
         return true
     }
+    
 }
