+++ ParticleEmitterDemo-SK/Images.xcassets/Spaceship.imageset/Contents.json
@@ -1,17 +1,21 @@
 {
   "images" : [
     {
+      "filename" : "Spaceship.png",
       "idiom" : "universal",
-      "scale" : "1x",
-      "filename" : "Spaceship.png"
+      "scale" : "1x"
     },
     {
       "idiom" : "universal",
       "scale" : "2x"
+    },
+    {
+      "idiom" : "universal",
+      "scale" : "3x"
     }
   ],
   "info" : {
-    "version" : 1,
-    "author" : "xcode"
+    "author" : "xcode",
+    "version" : 1
   }
 }
